@@ -11,7 +11,8 @@
     "build.types": "cross-env NODE_ENV=production tsc --project tsconfig.types.json",
     "build.test": "npm run build && npm run debug",
     "lint": "eslint --ext .js,.ts .",
-    "lint.fix": "eslint --ext .js,.ts --fix ."
+    "lint.fix": "eslint --ext .js,.ts --fix .",
+    "rollup-build": "rollup -cw"
   },
   "files": [
     "libs/"
@@ -36,14 +37,19 @@
     "@babel/preset-env": "^7.16.4",
     "@babel/preset-typescript": "^7.16.0",
     "@types/node": "^16.11.17",
+    "@rollup/plugin-node-resolve": "^13.0.6",
     "@types/read-pkg-up": "^6.0.0",
     "@types/webpack": "^5.28.0",
     "cross-env": "^7.0.3",
     "eslint": "^8.3.0",
     "read-pkg-up": "6.0.0",
+    "rollup": "^2.60.2",
+    "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-serve": "^1.1.0",
+    "rollup-plugin-typescript2": "^0.31.1",
     "typescript": "^4.5.2",
-    "webpack": "4.46.0",
-    "webpack-cli": "3.3.12"
+    "webpack": "^5.64.4",
+    "webpack-cli": "^4.9.1"
   },
   "peerDependencies": {
     "webpack": ">=4.0.0 <6.0.0"
